requirements.txt
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
authlib==1.3.2
google-ads>=25.0.0
